Changes to the hexRef4 class. Removed some functions and replaced with more general functions (setInternalFaceInfoA{0,1} -> setInternalFaceInfo)
Removed the setIndicator function as it is now unnecessary.

Added a section to findRelevantDir(..) to try and handle splits involving
pre-refined cells of different levels. This is a pretty bad fix and probably
will need to be improved/replaced later.

Current main concerns are with section 3 of face splitting in setRefinement.
The section is being used when a cell bordering already refined cells is later
refined. The pre-existing faces are handled by S3, and I think that the use
of just getFaceNeighbours(..) is causing problems.
At the very least, the refinement of these cells is coincident with the
occurance of errors which lead to maligned faces and connectivity issues.
This is also where the aspect * 3, 3/aspect section is used (bad fix described
above).

Changes made to the setting of cellDepth_ on construction - now have a
cell0Length_ field, which holds the length of an unrefined cell in each
direction as a Foam::vector. Consequently have reordered the constructor lists
to put the setting of this field before that of the aspectX_to_Y_ field, which
requires the values held by cell0Length_.