Changes to hexRef.C, now my mesh constructs without twists in a wider arrangement of cells / interface. Added a sortFaces class for diagnostics - this can be removed for a release library version, but is included here since hexRef4 now depends on the class. Added an "extended_debug" bool, which is static and accessible from solvers using hexRef4::extended_debug = 0, 1, true, false, etc. Some additional changes to comments to try and make slightly clearer in parts. Will now test the damBreak case sent to me earlier this week and try work out why that is breaking.
TODO - Part 3 of the face splitting only allows paired faces at present. This
is causing an error for me in my 51st time step for my test case, where
non-paired faces are somehow reaching part 3. Need to see if this is because
they're being missed, and should be handled earlier, or if I must relax the
restrictions on part 3 faces.